Reduce repo sync overhead
